{"schema":"decodex/commit/1","summary":"Split knowledge source snapshot modules","authority":"manual"}
